Latest upstream changes (#36)
* [cron testint] fix another deploy error (#4134)

Fixed another deploy to autopush error

This should fix things, ran cloud-build-local to test the yaml file

* [NL] Update timeline date range to work with FE change (#4111)

There was a FE change for how timeline tiles handle date ranges.
Timelines used to cut dates to the lowest granularity between specified
date range and observation date before comparing them, but now timeline
will only cut observation dates to the granularity of the specified date
range before comparing dates. On the NL side, we want the original
behavior so that if a start date like 2020-04 is specified, observations
from 2020 still show. This is a change to the date range we pass in the
tile config from the NL so that we get the original behavior.

Additional change:
- update the regex for date recognition because query will actually be
in lower case by the time we get to the date parsing code
- make sv_place_facet and sv_place_facets consistent to all be singular

* Added "Date of highest coverage" support to range slider and map web components (#4120)

* Updated slider web component to select HIGHEST_COVERAGE date
* Added `showTrendsSummary` option to `datacommons-slider` to show "Show
trends summary" checkbox.
* Added loading indicator to map tile
* Fixed issue where moving the slider caused map component to flash in
and out of view

## Showing date of highest coverage
![Screenshot 2024-04-17 at 3 17
16 PM](https://github.com/datacommonsorg/website/assets/13766/0c63671b-854c-45e5-8a37-b501051c07b4)

## Showing "trend summary"
![Screenshot 2024-04-18 at 4 45
04 PM](https://github.com/datacommonsorg/website/assets/13766/9948774b-55b4-4778-8ad0-2c1dfbe2823d)

## Showing other date
![Screenshot 2024-04-17 at 3 17
29 PM](https://github.com/datacommonsorg/website/assets/13766/c050ead8-34e8-4635-a40d-2cd589e11107)

## Loading chart for the first time
![Screenshot 2024-04-17 at 1 57
25 PM](https://github.com/datacommonsorg/website/assets/13766/507d0770-1665-4be9-bcaf-9c07c65f590d)

## Changing dates shows loading spinner
![Screenshot 2024-04-17 at 1 57
50 PM](https://github.com/datacommonsorg/website/assets/13766/d2918c76-1a25-4ab3-b525-707153c77700)

* Fix web compose docker build to use the tagged mixer and import submodule (#4136)

* NL: Merge Custom + Base indexes on the fly (#4132)

Builds on top of https://github.com/datacommonsorg/website/pull/4130

Currently, the custom_ft + medium_ft embeddings are merged on server
load.

Instead, we merge in custom_ft index in the query path, when the env is
custom DC (and the requested index is not custom_ft already).

Why?

* custom_ft can merge with any index besides medium_ft, based on the
`idx=` setting
* Allows interoperability between different custom DC vs. base DC
embedding stores. For example, custom DC might be in-memory while base
DC can be LanceDB or Vertex AI based.
* This will also allow unblocking UN work. So when we use `dc=undata`,
which implies `idx=undata_ft`, the stuff in `custom_ft` will get merged
in

* Disabled HIGHEST_COVERAGE option in SQL custom data commons (#4135)

Setting date=HIGHEST_COVERAGE falls back to date=LATEST in SQL custom
data commons

* Use anonymous GCS client only for custom dc and local user data paths. (#4137)

---------

Co-authored-by: chejennifer <69875368+chejennifer@users.noreply.github.com>
Co-authored-by: Dan Noble <dwnoble@google.com>
Co-authored-by: Bo Xu <shifucun@users.noreply.github.com>
Co-authored-by: Prashanth R <shanth@google.com>
Co-authored-by: Keyur Shah <keyurs@google.com>