Changes to the Data Commons core (with the intent that these later are also done in the Data Commons master in order not to break the fork). These changes are all done with the idea in mind of facilitating customization of the Data Commons application, in particular the various React applications and the Emotion theme.
To accomplish this, webpack (a core file) was updated to allow for the addition of an optional webpack custom file. Most repos will not include this (including the upstream master) but if included, the application will use this to augment the core webpack configuration. This allows custom forks, including this one, to use its own SCSS add its own react apps and override existing react apps to customize them.

Further customization was done to allow applications from adding their own topic pages using the configuration, rather than updating core code, and to allow a custom application to pass arbitrary environment variables through to the templates.